Validate product ID in update and delete endpoints; improve error handling

+++ Backend/controllers/product.controller.js
@@ -29,6 +29,7 @@ export const createProducts = async (req, res) => {
     export const updateProducts = async (req, res) => {
         const {id} = req.params
         const product = req.body
+
         if(!mongoose.Types.ObjectId.isValid(id)){
           return res.status(404).json({success:false, message:"Invalid product id/product not found"})
         }
@@ -42,10 +43,15 @@ export const createProducts = async (req, res) => {
 
     export const deleteProducts = async (req,res)=>{
     const{id} =req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+      return res.status(404).json({success:false, message:"Invalid product id/product not found"})
+    }
+    
     try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({success:true, message: "Product deleted"})  
     } catch (error) {
       console.log('Error in deleting product:', error.message);
-      res.status(404).json({success:false, message: "Product not found"})
+      res.status(500).json({success:false, message: "Server error"})  
     }}
